Implementation improvements
* Add TODO to support GPU memory

* Use singleton response allocator

* Move response iterators to _response.py

* Rename _allocators.py to _memorybuffer.py and cleanup

* Cleanup response_queue on request

* Raise exception when unable to move output tensor to requested memory type
