+++ python/tritonserver/_api/_request.py
@@ -34,7 +34,6 @@ from dataclasses import dataclass, field
 from typing import Any, Optional
 
 from tritonserver._api import _model
-from tritonserver._api._allocators import MemoryAllocator
 from tritonserver._api._datautils import CustomKeyErrorDict
 from tritonserver._api._dlpack import DLDeviceType as DLDeviceType
 from tritonserver._api._tensor import Tensor
@@ -86,18 +85,6 @@ class InferenceRequest:
         fallback. Memory type can be given as a string, MemoryType,
         tuple [MemoryType, memory_type__id], or tuple[DLDeviceType,
         device_id].
-    output_memory_allocator : Optional[MemoryAllocator], default None
-        Memory allocator to use for inference response output. If not
-        provided default allocators will be used to allocate
-        MemoryType.GPU or MemoryType.CPU memory as set in
-        output_memory_type or as requested by backend / model
-        preference.
-    response_queue : Optional[Union[queue.SimpleQueue, asyncio.Queue]], default None
-        Queue for asynchronous handling of inference responses. If
-        provided Inference responses will be added to the queue in
-        addition to the response iterator. Must be queue.SimpleQueue
-        for non asyncio requests and asyncio.Queue for asyncio
-        requests.
 
     Examples
     --------
@@ -132,8 +119,6 @@ class InferenceRequest:
     inputs: dict[str, Tensor | Any] = field(default_factory=dict)
     parameters: dict[str, str | int | bool | float] = field(default_factory=dict)
     output_memory_type: Optional[DeviceOrMemoryType] = None
-    output_memory_allocator: Optional[MemoryAllocator] = None
-    response_queue: Optional[queue.SimpleQueue | asyncio.Queue] = None
     _serialized_inputs: dict[str, Tensor] = field(init=False, default_factory=dict)
     _server: TRITONSERVER_Server = field(init=False)
